RDRealty/requirements.txt
asgiref==3.11.0
Django==6.0
mysqlclient==2.2.7
PyMySQL==1.1.2
sqlparse==0.5.4
tzdata==2025.2
google-generativeai==0.3.2
python-dotenv==1.0.1
Pillow
